+++ cmd/aem/instance.go
@@ -226,126 +226,3 @@ func (c *CLI) instanceListCmd() *cobra.Command {
 		},
 	}
 }
-
-func (c *CLI) instanceBackupCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "backup",
-		Aliases: []string{"bak"},
-		Short:   "Manages AEM instance backups",
-	}
-	cmd.AddCommand(c.instanceBackupMakeCmd())
-	cmd.AddCommand(c.instanceBackupListCmd())
-	cmd.AddCommand(c.instanceBackupUseCmd())
-	return cmd
-}
-
-func (c *CLI) instanceBackupMakeCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "make",
-		Aliases: []string{"mk", "create"},
-		Short:   "Makes AEM instance backup",
-		Run: func(cmd *cobra.Command, args []string) {
-			instanceManager := c.aem.InstanceManager()
-			localInstance, err := instanceManager.OneLocal()
-			if err != nil {
-				c.Error(err)
-				return
-			}
-			instance := localInstance.Instance()
-
-			file, _ := cmd.Flags().GetString("file")
-			if file == "" {
-				file = localInstance.ProposeBackupFile()
-			}
-
-			running := localInstance.IsRunning()
-			if running {
-				if localInstance.StopAndAwait(); err != nil {
-					c.Error(err)
-					return
-				}
-			}
-			if err := localInstance.MakeBackup(file); err != nil {
-				c.Error(err)
-				return
-			}
-			if running {
-				if err := localInstance.StartAndAwait(); err != nil {
-					c.Error(err)
-					return
-				}
-			}
-
-			c.SetOutput("instance", instance)
-			c.SetOutput("file", file)
-			c.Ok("instance backup made")
-		},
-	}
-	cmd.Flags().String("file", "", "Local file path")
-	return cmd
-}
-
-func (c *CLI) instanceBackupListCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "list",
-		Aliases: []string{"ls"},
-		Short:   "Lists available AEM instance backups",
-		Run: func(cmd *cobra.Command, args []string) {
-			backups, err := c.aem.InstanceManager().ListBackups()
-			if err != nil {
-				c.Error(err)
-				return
-			}
-			c.SetOutput("backups", backups)
-			c.Ok("instance backup listed")
-		},
-	}
-	return cmd
-}
-
-func (c *CLI) instanceBackupUseCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "use",
-		Aliases: []string{"read", "load"},
-		Short:   "Use AEM instance backup",
-		Run: func(cmd *cobra.Command, args []string) {
-			instanceManager := c.aem.InstanceManager()
-			localInstance, err := instanceManager.OneLocal()
-			if err != nil {
-				c.Error(err)
-				return
-			}
-			deleteCreated, _ := cmd.Flags().GetBool("delete-created")
-			if !deleteCreated && localInstance.IsCreated() {
-				c.Fail("instance already created; to delete it use flag '--delete-created'")
-				return
-			}
-			instance := localInstance.Instance()
-			file, _ := cmd.Flags().GetString("file")
-			running := localInstance.IsRunning()
-			if running {
-				if localInstance.StopAndAwait(); err != nil {
-					c.Error(err)
-					return
-				}
-			}
-			if err := localInstance.UseBackup(file, deleteCreated); err != nil {
-				c.Error(err)
-				return
-			}
-			if running {
-				if err := localInstance.StartAndAwait(); err != nil {
-					c.Error(err)
-					return
-				}
-			}
-			c.SetOutput("instance", instance)
-			c.SetOutput("file", file)
-			c.Ok("instance backup used")
-		},
-	}
-	cmd.Flags().String("file", "", "Local file path")
-	cmd.MarkFlagRequired("file")
-	cmd.Flags().Bool("delete-created", false, "Delete already created instance")
-	return cmd
-}
